pyproject.toml
[tool.poetry]
name = "glider-dash"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dash = "^2.14.1"
polars = "^0.19.16"
pandas = "^2.1.3"
dash-bootstrap-components = "^1.5.0"
tqdm = "^4.66.1"
geographiclib = "^2.0"
pygeodesy = "^23.11.11"
pyarrow = "^14.0.1"
dash-uploader = "^0.6.0"
packaging = "21.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"